refactor: rename shoppingCart package to shopping_cart for consistency
