+++ artists-heaven-backend/src/main/java/com/artists_heaven/entities/user/User.java
@@ -7,7 +7,7 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import com.artists_heaven.shoppingCart.ShoppingCart;
+import com.artists_heaven.shopping_cart.ShoppingCart;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import jakarta.persistence.CascadeType;
