+++ artists-heaven-backend/src/main/java/com/artists_heaven/shopping_cart/ShoppingCartRepository.java
@@ -1,4 +1,4 @@
-package com.artists_heaven.shoppingCart;
+package com.artists_heaven.shopping_cart;
 
 import java.util.Optional;
